+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Fixed
 - fix `runtime.isDroppedFileFromOutside` is not working as indended #5165
 - tauri: fix drag and drop on macOS #5165
+- translate "Emoji" and "Sticker" in emoji & sticker picker
+- tauri: save zoom level between webxdc app launches #5163
 
 <a id="1_59_1"></a>